+++ BenchTests/BenchTests.swift
@@ -52,7 +52,6 @@ class BenchTests: XCTestCase, CLLocationManagerDelegate {
     func navigationViewController(route: Route, locationManager: ReplayLocationManager) -> NavigationViewController {
         
         let speechAPI = SpeechAPISpy(accessToken: token)
-        let voiceController = MapboxVoiceController(speechClient: speechAPI, audioPlayerType: AudioPlayerDummy.self)
         let directions = DirectionsSpy(accessToken: token)
         let service = MapboxNavigationService(route: route,
                                               directions: directions,
@@ -60,8 +59,11 @@ class BenchTests: XCTestCase, CLLocationManagerDelegate {
                                               eventsManagerType: NavigationEventsManagerSpy.self,
                                               simulating: .never,
                                               routerType: RouteController.self)
+        let voiceController = MapboxVoiceController(navigationService: service, speechClient: speechAPI, audioPlayerType: AudioPlayerDummy.self)
         
-        return NavigationViewController(for: route, navigationService: service, voiceController: voiceController)
+        let options = NavigationOptions(navigationService: service, voiceController: voiceController)
+        
+        return NavigationViewController(for: route, options: options)
     }
 }