+++ MapboxCoreNavigationTests/MapboxCoreNavigationTests.swift
@@ -224,6 +224,103 @@ class MapboxCoreNavigationTests: XCTestCase {
         }
     }
     
+    func testOrderOfExecution() {
+        route.accessToken = "foo"
+        
+        let trace = Fixture.generateTrace(for: route).shiftedToPresent().qualified()
+        navigation = MapboxNavigationService(route: route)
+        
+        struct InstructionPoint {
+            enum InstructionType {
+                case visual, spoken
+            }
+            
+            let type: InstructionType
+            let legIndex: Int
+            let stepIndex: Int
+            let spokenInstructionIndex: Int
+            let visualInstructionIndex: Int
+        }
+        
+        var points = [InstructionPoint]()
+        
+        expectation(forNotification: .routeControllerDidPassSpokenInstructionPoint, object: nil) { (notification) -> Bool in
+            let routeProgress = notification.userInfo![RouteControllerNotificationUserInfoKey.routeProgressKey] as! RouteProgress
+            let legIndex = routeProgress.legIndex
+            let stepIndex = routeProgress.currentLegProgress.stepIndex
+            let spokenInstructionIndex = routeProgress.currentLegProgress.currentStepProgress.spokenInstructionIndex
+            let visualInstructionIndex = routeProgress.currentLegProgress.currentStepProgress.visualInstructionIndex
+            
+            let point = InstructionPoint(type: .spoken, legIndex: legIndex, stepIndex: stepIndex, spokenInstructionIndex: spokenInstructionIndex, visualInstructionIndex: visualInstructionIndex)
+            points.append(point)
+            
+            return true
+        }
+        
+        expectation(forNotification: .routeControllerDidPassVisualInstructionPoint, object: nil) { (notification) -> Bool in
+            let routeProgress = notification.userInfo![RouteControllerNotificationUserInfoKey.routeProgressKey] as! RouteProgress
+            let legIndex = routeProgress.legIndex
+            let stepIndex = routeProgress.currentLegProgress.stepIndex
+            let spokenInstructionIndex = routeProgress.currentLegProgress.currentStepProgress.spokenInstructionIndex
+            let visualInstructionIndex = routeProgress.currentLegProgress.currentStepProgress.visualInstructionIndex
+            
+            let point = InstructionPoint(type: .visual, legIndex: legIndex, stepIndex: stepIndex, spokenInstructionIndex: spokenInstructionIndex, visualInstructionIndex: visualInstructionIndex)
+            points.append(point)
+            
+            return true
+        }
+        
+        for location in trace {
+            navigation.router!.locationManager!(navigation.locationManager, didUpdateLocations: [location])
+        }
+        
+        waitForExpectations(timeout: 2) { (error) in
+            XCTAssertNil(error)
+        }
+        
+        XCTAssertEqual(points[0].legIndex, 0)
+        XCTAssertEqual(points[0].stepIndex, 0)
+        XCTAssertEqual(points[0].visualInstructionIndex, 0)
+        XCTAssertEqual(points[0].spokenInstructionIndex, 0)
+        XCTAssertEqual(points[0].type, .spoken)
+        
+        XCTAssertEqual(points[1].legIndex, 0)
+        XCTAssertEqual(points[1].stepIndex, 0)
+        XCTAssertEqual(points[1].visualInstructionIndex, 0)
+        XCTAssertEqual(points[1].spokenInstructionIndex, 0)
+        XCTAssertEqual(points[1].type, .visual)
+        
+        XCTAssertEqual(points[2].legIndex, 0)
+        XCTAssertEqual(points[2].stepIndex, 0)
+        XCTAssertEqual(points[2].visualInstructionIndex, 0)
+        XCTAssertEqual(points[2].spokenInstructionIndex, 1)
+        XCTAssertEqual(points[2].type, .spoken)
+        
+        XCTAssertEqual(points[3].legIndex, 0)
+        XCTAssertEqual(points[3].stepIndex, 1)
+        XCTAssertEqual(points[3].visualInstructionIndex, 0)
+        XCTAssertEqual(points[3].spokenInstructionIndex, 0)
+        XCTAssertEqual(points[3].type, .spoken)
+        
+        // Make sure we never have unsynced indexes or move backward in time by comparing previous to current instruction point
+        let zippedPoints = zip(points, points.suffix(from: 1))
+        
+        for seq in zippedPoints {
+            let previous = seq.0
+            let current = seq.1
+            
+            let sameStepAndLeg = previous.legIndex == current.legIndex && previous.stepIndex == current.stepIndex
+            
+            if sameStepAndLeg {
+                XCTAssert(current.visualInstructionIndex >= previous.visualInstructionIndex)
+                XCTAssert(current.spokenInstructionIndex >= previous.spokenInstructionIndex)
+            } else {
+                XCTAssert(current.visualInstructionIndex == 0)
+                XCTAssert(current.spokenInstructionIndex == 0)
+            }
+        }
+    }
+    
     func testFailToReroute() {
         enum TestError: Error {
             case test